Move ChangeShapeTypeOption out of ModelTransformer
